Mapvote123 and AutoComplete working (Filters not implemented yet)
