+++ squad-server/package.json
@@ -16,7 +16,7 @@
     ".": "./index.js",
     "./logo": "./utils/print-logo.js",
     "./plugins": "./plugins/index.js",
-    "./plugins/base": ".plugins/base-plugin.js",
+    "./plugins/base": "./plugins/base-plugin.js",
     "./server-events": "./core/server-events.js",
     "./constants": "./utils/constants.js"
   }
