+++ squad-server/plugins/index.js
@@ -11,6 +11,8 @@ import DiscordRoundWinner from './discord-round-winner.js';
 import DiscordServerStatus from './discord-server-status.js';
 import DiscordSubsystemRestarter from './discord-subsystem-restarter.js';
 import IntervalledBroadcasts from './intervalled-broadcasts.js';
+import MapVote123 from './mapvote/mapvote-123.js';
+import MapVoteAutoComplete from './mapvote/mapvote-auto-complete.js';
 import SeedingMode from './seeding-mode.js';
 import TeamRandomizer from './team-randomizer.js';
 
@@ -29,7 +31,9 @@ const plugins = [
   DiscordSubsystemRestarter,
   IntervalledBroadcasts,
   SeedingMode,
-  TeamRandomizer
+  TeamRandomizer,
+  MapVoteAutoComplete,
+  MapVote123
 ];
 
 const pluginsByName = {};
