+++ squad-server/utils/squad-layers.js
@@ -4,7 +4,7 @@ import fs from 'fs';
 
 class SquadLayersBase {
   get layerNames() {
-    return this.layers.map((layer) => layer.name);
+    return this.layers.map((layer) => layer.layer);
   }
 
   getLayerByCondition(condition) {
@@ -21,7 +21,7 @@ class SquadLayersBase {
   }
 
   getLayerByLayerNameAutoCorrection(layerName) {
-    return this.getLayerByLayerName(didYouMean(layerName, this.layerNames()));
+    return this.getLayerByLayerName(didYouMean(layerName, this.layerNames));
   }
 
   getLayerByNumber(layerNumber) {
@@ -46,10 +46,14 @@ class SquadLayers extends SquadLayersBase {
   }
 
   buildPoolFromLayerNames(layerNames, activeFilter) {
-    return new SquadLayersPool(
-      this.layers.filter((layer) => layerNames.includes(layer.layer)),
-      activeFilter
-    );
+    // var layers =  this.layers.filter((layer) => layerNames.includes(layer.layer));
+    var layers = this.layers.filter((layer) => {
+      return layerNames.some((filteredLayer) => {
+        return layer.layer === filteredLayer.layer;
+      });
+    });
+
+    return new SquadLayersPool(layers, activeFilter);
   }
 
   buildPoolFromLayerNamesAutoCorrection(layerNames, activeFilter) {
